fix(core): declare direct libc dependency for O_NOFOLLOW sidecar opens
mmap_store opens cold-erasure sidecars and journals with
libc::O_NOFOLLOW but never declared libc as a direct unix dependency;
it only compiled through a transitive edge. Declare it explicitly so
the durable-scrollback storage path no longer depends on another
crate's dependency tree.

While here, compare shard.index to the loop index as u8 on both sides:
recover_from_erasure_sidecars already converts the index up front via
u8::try_from, so the usize::from round-trip was redundant.

Caveat: Cargo.lock is untouched because libc is already locked.

+++ crates/frankenterm-core/src/storage/mmap_store.rs
@@ -1245,7 +1245,7 @@ impl PaneFile {
             }
             Self::revalidate_open_file(&path, &file)?;
             match ColdErasureShard::decode_bytes(&bytes) {
-                Ok(shard) if usize::from(shard.index) == index => shards.push(shard),
+                Ok(shard) if shard.index == index => shards.push(shard),
                 Ok(shard) => invalid_shards.push(format!(
                     "path index {index} contains shard index {}",
                     shard.index
